Use generic ArgumentError for invalid argument
We have not only "type" argument but also "n_classes" argument. It's
better that we use generic ArgumentError for invalid argument instead
of using each specific exception class.

Another idea: We create exception classes for each argument as a
subclass of ArgumentError. But it seems that it's too much
work.

+++ test/helper.rb
@@ -13,6 +13,7 @@ module Helper
     end
 
     def teardown_sandbox
+      return unless defined?(@tmp_dir)
       FileUtils.rm_rf(@tmp_dir)
     end
   end
